Add house and center of gravity for camera location
This is midnight programming, I want to rewrite quite a bit. Move stuff
around. Currently it centers the camera on the entire world. In the
future it should only be things in a certain distance from the camera.
New objects with gravity that come into range should accelerate the
camera towards the new center of gravity. This is todo naturally.